Move component, make a test
