+++ hooks/hooks
@@ -1,0 +1,1 @@
+./.git/hooks/
