+++ test/claims-status/components/appeals-v2/AppealListItemV2.unit.spec.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { expect } from 'chai';
+import AppealListItemV2 from '../../../../src/js/claims-status/components/appeals-v2';
+
+describe('<AppealListItemV2/>', () => {
+
+});
